backend/requirements.txt
# aiohappyeyeballs==2.6.1
# aiohttp==3.13.3
# aiosignal==1.4.0
# annotated-doc==0.0.4
# annotated-types==0.7.0
# anyio==4.12.1
# attrs==25.4.0
# certifi==2026.2.25
# charset-normalizer==3.4.4
# click==8.3.1
# dataclasses-json==0.6.7
# faiss-cpu==1.13.2
# fastapi==0.133.1
# filelock==3.24.3
# frozenlist==1.8.0
# fsspec==2026.2.0
# h11==0.16.0
# hf-xet==1.3.1
# httpcore==1.0.9
# httpx==0.28.1
# huggingface_hub==1.5.0
# idna==3.11
# Jinja2==3.1.6
# joblib==1.5.3
# jsonpatch==1.33
# jsonpointer==3.0.0
# langchain==0.1.20
# langchain-community==0.0.38
# langchain-core==0.1.53
# langchain-huggingface==0.0.3
# langchain-text-splitters==0.0.2
# langsmith==0.1.147
# markdown-it-py==4.0.0
# MarkupSafe==3.0.3
# marshmallow==3.26.2
# mdurl==0.1.2
# mpmath==1.3.0
# multidict==6.7.1
# mypy_extensions==1.1.0
# orjson==3.11.7
# packaging==23.2
# propcache==0.4.1
# pydantic==2.12.5
# pydantic_core==2.41.5
# Pygments==2.19.2
# pypdf==6.7.4
# python-dotenv==1.2.2
# python-multipart==0.0.22
# PyYAML==6.0.3
# regex==2026.2.19
# requests==2.32.5
# requests-toolbelt==1.0.0
# rich==14.3.3
# safetensors==0.7.0
# shellingham==1.5.4
# SQLAlchemy==2.0.47
# starlette==0.52.1
# tenacity==8.5.0
# threadpoolctl==3.6.0
# tqdm==4.67.3
# typer==0.24.1
# typer-slim==0.24.0
# typing-inspect==0.9.0
# typing-inspection==0.4.2
# typing_extensions==4.15.0
# urllib3==2.6.3
# uvicorn==0.41.0
# yarl==1.22.0
fastapi==0.133.1
uvicorn==0.41.0
langchain==0.1.20
langchain-core==0.1.53
langchain-community==0.0.38
langchain-huggingface==0.0.3
huggingface_hub==1.5.0
pypdf==6.7.4
python-dotenv==1.2.2
python-multipart==0.0.22
faiss-cpu==1.13.2
httpx==0.28.1
requests==2.32.5
pydantic==2.12.5
starlette==0.52.1